refactor(ui): Use barrel export also for form subfields
Also barrel-export the Advisor, Analyzer etc. fields from their
corresponding `-components` directory, which simplifies their import
especially in the ORT run creation's main form.

This refactor brings the UI importing structure more aligned:
- for the most part, Vite's `@` path alias is used in imports
- as an exception, absolute import paths are used when the files are
  closely co-located, such as in different `-component` directories
